+++ src/main/java/kitchenpos/tablegroup/application/dto/TableGroupCreateRequestDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.tablegroup.application.dto;
 
 import java.util.List;
