+++ src/main/java/kitchenpos/tablegroup/model/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.model;
 
 import java.time.LocalDateTime;
