+++ src/main/java/kitchenpos/tablegroup/model/TableGroupVerifier.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.model;
 
 import java.util.List;
