+++ src/main/java/kitchenpos/tablegroup/repository/TableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.repository;
+package kitchenpos.tablegroup.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.model.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
