+++ src/test/java/kitchenpos/tablegroup/model/TableGroupVerifierTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.model;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,8 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import kitchenpos.tablegroup.model.TableGroupVerifier;
 
 class TableGroupVerifierTest {
